moved GlobalDebuggable out of base folder
