+++ src/main/java/frc/robot/utils/GlobalDebuggable.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.utils;
 
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
